requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
edge-tts
moviepy==1.0.3
numpy